+++ src/map/map-parser.cpp
@@ -33,10 +33,10 @@ MapParser::parse(std::string path, Grid & map)
 	std::string line;
 	char c;
 	std::vector<char> *current_line;
-	for(unsigned y(0) ; std::getline(file, line) && y < map.size ; ++y)
+	for(int y(0) ; std::getline(file, line) && y < map.size ; ++y)
 	{
 		current_line = new std::vector<char>();
-		for(unsigned x(0) ; x < map.size ; ++x)
+		for(int x(0) ; x < map.size ; ++x)
 		{
 			try
 			{
@@ -62,5 +62,5 @@ MapParser::parse(std::string path, Grid & map)
 		delete(current_line);
 	}
 
-	return (map.grid.size() == map.size);
+	return (map.grid.size() == static_cast<unsigned>(map.size));
 }
